% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.comparison.R
\name{calcVst}
\alias{calcVst}
\title{Vst
Estimate the Vst among two populations from a gds previouly loaded with LRR (with \link{importLRR_BAF} function)}
\usage{
calcVst(all.paths, pops, chr.code.name = NULL, n.cor = 1)
}
\arguments{
\item{all.paths}{Object returned from \code{CreateFolderTree} function with the working folder tree}

\item{pops}{Two populations to be compared}

\item{chr.code.name}{A data-frame with the integer name in the first column and the original name for each chromosome}

\item{n.cor}{Number of cores to be used}
}
\value{
Vst for all probes with CNV genotypes listed in the gds
}
\description{
Vst
Estimate the Vst among two populations from a gds previouly loaded with LRR (with \link{importLRR_BAF} function)
}
