% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_assoc.R
\name{cnvExprAssoc}
\alias{cnvExprAssoc}
\title{CNV-expression association analysis}
\usage{
cnvExprAssoc(cnvrs, calls, rcounts, window = "1Mbp",
  multi.calls = "largest", min.samples = 10, min.cpm = 2,
  de.method = c("edgeR", "limma"), padj.method = "BH",
  verbose = FALSE)
}
\arguments{
\item{cnvrs}{A \code{\linkS4class{GRanges}} object containing the summarized
CNV regions as e.g. obtained with \code{\link{populationRanges}}.}

\item{calls}{A \code{\linkS4class{GRangesList}} or a 
\code{linkS4class{RaggedExperiment}} storing the individual CNV calls for 
each sample.}

\item{rcounts}{A \code{\linkS4class{RangedSummarizedExperiment}} storing the 
raw RNA-seq read counts in a rectangular fashion (genes x samples).}

\item{window}{Defaults to \code{"1Mbp"}.}

\item{multi.calls}{Character or a function. Defaults to \code{"largest"}.}

\item{min.samples}{Integer. Defaults to 10.}

\item{min.cpm}{Integer. Should genes not satisfying a minimum counts-per-million 
(cpm) threshold be excluded from the analysis? This is typically recommended. 
See the edgeR vignette for details. The default filter is to exclude genes 
with cpm < 2 in more than half of the samples.}

\item{de.method}{Character. Differential expression method. 
Defaults to \code{"edgeR"}.}

\item{padj.method}{Character. Method for adjusting p-values to multiple testing.
For available methods see the man page of the function \code{\link{p.adjust}}. 
Defaults to \code{"BH"}.}

\item{verbose}{Logical. Display progress messages? Defaults to \code{FALSE}.}
}
\value{
TODO
}
\description{
Testing CNV regions for effects on the expression level of genes in defined
genomic windows.  
 
Details here.
}
\examples{

states <- sample(c(0,1,3,4), 17, replace=TRUE)

calls <- GRangesList(
     sample1 = GRanges( c("chr1:1-10", "chr2:15-18", "chr2:25-34"), state=states[1:3]),
     sample2 = GRanges( c("chr1:1-10", "chr2:11-18" , "chr2:25-36"), state=states[4:6] ),
     sample3 = GRanges( c("chr1:2-11", "chr2:14-18", "chr2:26-36"), state=states[7:9] ),
     sample4 = GRanges( c("chr1:1-12", "chr2:18-35" ), state=states[10:11] ),
     sample5 = GRanges( c("chr1:1-12", "chr2:11-17" , "chr2:26-34"), state=states[12:14] ) ,
     sample6 = GRanges( c("chr1:1-12", "chr2:12-18" , "chr2:25-35"), state=states[15:17] )
) 

cnvrs <- populationRanges(calls, density=0.1)



}
\references{
Geistlinger et al. (2018) Widespread modulation of gene expression
by copy number variation in skeletal muscle. Sci Rep, 8(1):1399.
}
\seealso{
\code{\link{findOverlaps}} to find overlaps between sets of genomic 
regions, \code{\link{qreduceAssay}} to summarize ragged genomic location data 
in defined genomic regions, \code{\link{glmQLFit}} and \code{\link{glmQLFTest}}
to conduct negative binomial generalized linear models for RNA-seq read 
count data.
}
\author{
Ludwig Geistlinger <Ludwig.Geistlinger@sph.cuny.edu>
}
