% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phen_assoc_epistasis.R
\name{cnvGWAS.V2}
\alias{cnvGWAS.V2}
\title{Run the CNV-GWAS}
\usage{
cnvGWAS.V2(phen.info, n.cor = 1, min.sim = 0.95, freq.cn = 0.01,
  snp.matrix = FALSE, method.m.test = "fdr", lo.phe = 1,
  chr.code.name = NULL, genotype.nodes = "CNVGenotype",
  coding.translate = "all", path.files = NULL, list.of.files = NULL,
  produce.gds = TRUE, run.lrr = FALSE, assign.probe = "min.pvalue",
  select.population = NULL, epistasis.mode = FALSE,
  correct.inflation = FALSE, both.up.down = FALSE,
  cnvr.specific = FALSE, exclude.outliers = FALSE, minq = 0.01,
  maxq = 0.99, log.pheno = FALSE, verbose = FALSE)
}
\arguments{
\item{phen.info}{Returned by \code{\link{setupCnvGWAS}}}

\item{n.cor}{Number of cores to be used}

\item{min.sim}{Minimum CNV genotype distribuition similarity among subsequent probes. Default is 0.95 (i.e. 95\%)}

\item{freq.cn}{Minimum CNV frequency where 1 (i.e. 100\%), or all samples deviating from diploid state. Default 0.01 (i.e. 1\%)}

\item{snp.matrix}{Only FALSE implemented - If TRUE B allele frequencies (BAF) would be used to reconstruct CNV-SNP genotypes}

\item{method.m.test}{Correction for multiple tests to be used. FDR is default, see \code{\link{(p.adjust)}} for
other methods.}

\item{lo.phe}{The phenotype to be analyzed in the PhenInfo$phenotypesSam data-frame}

\item{chr.code.name}{A data-frame with the integer name in the first column and the original name for each chromosome}

\item{genotype.nodes}{Expression data type. Nodes with CNV genotypes to be produced in the gds file.}

\item{coding.translate}{For 'CNVgenotypeSNPlike'. If NULL or unrecognized string use only biallelic CNVs. If 'all' code multiallelic CNVs as 0 
for loss; 1 for 2n and 2 for gain.}

\item{path.files}{Folder containing the input CNV files used for the CNV calling (i.e. one text file with 5 collumns 
for each sample). Columns should contain (i) probe name, (ii) Chromosome, (iii) Position, (iv) LRR and (v) BAF. Only if \sQuote{lrr} == TRUE}

\item{list.of.files}{Data-frame with two columns where the (i) is the file name with signals and (ii) is the 
correspondent name of the sample in the gds file. Only if \sQuote{lrr} == TRUE}

\item{produce.gds}{logical. If TRUE produce a new gds, if FALSE use gds previously created}

\item{run.lrr}{If TRUE use LRR values instead absolute copy numbers in the association}

\item{assign.probe}{\sQuote{min.pvalue} or \sQuote{high.freq} to represent the CNV segment}

\item{select.population}{Select a subset of populations to be used in the GWAS. Use the population name or names}

\item{epistasis.mode}{logical. In addition, perform the GWAS analysis using pairwise combination of the genotypes of all CNV segments to
understand epistatic effect among CNVs. Default is FALSE.}

\item{both.up.down}{logical. Probe genotype similarity considering the up- and downstream vicinity probes. If FALSE will only
downstream probes are considered}

\item{cnvr.specific}{Compare probes inside the same CNVR. Recommended to run with \sQuote{min.sim}>1}

\item{exclude.outliers}{logical. Use quadrants to exclude samples with extreme phenotypes (i.e. outliers)}

\item{minq}{Minimum quadrant to exclude. Only if \sQuote{exclude.outliers} is TRUE}

\item{log.pheno}{logical. log2 transform the phenotypes}

\item{verbose}{Show progress in the analysis}

\item{lrr}{logical. Import LRR/BAF values from external files}

\item{minq}{Maximum quadrant to exclude. Only if \sQuote{exclude.outliers} is TRUE}
}
\value{
The CNV segments and the representative probes with their respective assigned p-value
}
\description{
Wraps all the necessary functions to run a CNV-GWAS using the output of the \code{\link{setupCnvGWAS}} function
 
(i) Produces the GDS file containing the genotype information  (if produce.gds == TRUE),
(ii) Produces the requested inputs for a linear (PLINK, http://zzz.bwh.harvard.edu/plink/gvar.shtml) or linear mixed 
model analysis, (iii) run a CNV-GWAS analysis and (iv) export a QQ-plot displaying the adjusted p-values. In this release, 
SNP genotypes are not considered and only the p-value for the copy number is available  (i.e. 'P(CNP)' in PLINK).
}
\examples{

# Load phenotype-CNV information

folder.package <- system.file('extdata', package = 'cnvAnalyzeR')

phen.info <- setupCnvGWAS(name="Example", phen.loc=file.path(folder.package, "Pheno.txt"), 
cnv.out.loc = file.path(folder.package, "CNVOut.txt"), 
map.loc = file.path(folder.package, "MapPenn.txt"))

# Define chr correspondence to numeric, if necessary
df <- "16 1A
25 4A
29 25LG1
30 25LG2
31 LGE22"

chr.code.name <- read.table(text=df, header=F)

segs.pvalue.gr <- cnvGWAS.V2(phen.info, chr.code.name=chr.code.name)
 
}
\author{
Vinicius Henrique da Silva <vinicius.dasilva@wur.nl>
}
