% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.comparison.R
\name{compareCNVsegFreq}
\alias{compareCNVsegFreq}
\title{HELPER - Admixture - TODO}
\usage{
compareCNVsegFreq(all.paths, segs.pvalue.gr, pops, p.value.min = 0.05,
  n.cor = 1)
}
\arguments{
\item{all.paths}{Object returned from \code{CreateFolderTree} function with the working folder tree}

\item{segs.pvalue.gr}{Object returned from \code{\link{setupCnvGWAS}} function}

\item{pops}{Two populations to be compared}

\item{p.value.min}{Minimum p-value to consider normal distribution}

\item{n.cor}{Number of cores to be used}
}
\description{
Compare CNV segment frequencies among populations and draw Fst for the CNVgenotypeSNPlike node
}
