% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.comparison.R
\name{PCAcnv}
\alias{PCAcnv}
\title{PCA with CNVs}
\usage{
PCAcnv(
  all.paths,
  segs.pvalue.gr,
  n.cor = 1,
  pops.names = NULL,
  method.ld = "dprime",
  attach.pca = TRUE,
  use.seg = TRUE
)
}
\arguments{
\item{all.paths}{Object returned from \code{CreateFolderTree} function with 
the working folder tree}

\item{segs.pvalue.gr}{Result from \code{\link{cnvGWAS}}}

\item{n.cor}{Number of cores}

\item{pops.names}{Name of the populations}

\item{method.ld}{LD method, "r2" or "dprime"}

\item{attach.pca}{Create a node in the CNV.gds containing the PCA analysis}
}
\description{
PCA with CNVs
}
\examples{

}
\author{
Vinicius Henrique da Silva <vinicius.dasilva@wur.nl>
}
