% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_ranges.R
\name{populationRangesRO}
\alias{populationRangesRO}
\title{Summarizing genomic ranges based on reciprocal overlap across a population}
\usage{
populationRangesRO(grl, ro.thresh = 0.5, multi.assign = FALSE)
}
\arguments{
\item{grl}{A \code{\linkS4class{GRangesList}}.}

\item{ro.thresh}{Numeric. Threshold for reciprocal overlap required for merging
two overlapping regions. Defaults to 0.5.}

\item{multi.assign}{Logical. Allow regions to be assigned to several region clusters?
Defaults to FALSE.}
}
\value{
A \code{\linkS4class{GRanges}} object containing the summarized
genomic ranges.
}
\description{
Reciprocal overlap (RO) approach (e.g. Conrad et al., Nature, 2010)
}
\details{
Reciprocal overlap of 0.51 between two genomic regions A and B:

requires that B overlap at least 51% of A, 
  *and* that A also overlaps at least 51% of B

Approach:

At the top level of the hierarchy, all contiguous bases overlapping at least 
1bp of indvidual calls are merged into one region. 
Within each region, we further define reciprocally overlapping regions with 
the following algorithm:

\itemize{
\item Calculate reciprocal overlap (RO) between all remaining calls.
\item Identify pair of calls with greatest RO. If RO > threshold, merge and 
      create a new CNV. If not, exit.
\item Continue adding unclustered calls to the region, in order of best overlap. 
      In order to add a call, the new call must have > threshold to all calls 
      within the region to be added. When no additional calls may be added, 
      move to next step.
\item If calls remain, return to 1. Otherwise exit.
}
}
\examples{

grl <- GRangesList(
     sample1 = GRanges( c("chr1:1-10", "chr2:15-18", "chr2:25-34") ),
     sample2 = GRanges( c("chr1:1-10", "chr2:11-18" , "chr2:25-36") ),
     sample3 = GRanges( c("chr1:2-11", "chr2:14-18", "chr2:26-36") ),
     sample4 = GRanges( c("chr1:1-12", "chr2:18-35" ) ),
     sample5 = GRanges( c("chr1:1-12", "chr2:11-17" , "chr2:26-34") ) ,
     sample6 = GRanges( c("chr1:1-12", "chr2:12-18" , "chr2:25-35") )
)

# default as chosen in the original CNVRuler procedure
populationRangesRO(grl, ro.thresh=0.5)

}
\references{
Conrad et al. (2010) Origins and functional impact of copy number variation 
in the human genome. Nature, 464(7289):704-12.
}
\seealso{
\code{\link{subsetByOverlaps}}
}
\author{
Ludwig Geistlinger
}
