% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_assoc.R
\name{produceRandomPheno}
\alias{produceRandomPheno}
\title{Produce list simulation to be used in the cnvGWAS()}
\usage{
produceRandomPheno(simulation, model, phen.info)
}
\arguments{
\item{simulation}{Numeric or NULL. If numeric, simulated phenotypes will be used. The number
will define the number of permutation (i.e. if simualtion = 1000, the analysis will be performed
1000 times with random phenotypes). The random phenotypes have the same standard deviation and mean
of the real phenotype defined in the model.}

\item{model}{Linear mixed model to fit (without the CNV). Use \code{\link{formula}}}

\item{phen.info}{Returned by \code{\link{setupCnvGWAS}}}
}
\description{
Produce list simulation to be used in the cnvGWAS()
}
