% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_assoc.R
\name{snpgdsGetGenoCNV}
\alias{snpgdsGetGenoCNV}
\title{Extract the CNV genotypes from the GDS file}
\usage{
snpgdsGetGenoCNV(genofile, snp.id, node.to.extract = "CNVgenotype")
}
\arguments{
\item{genofile}{is the loaded gds file}

\item{snp.id}{is the SNP probe name to retrieve}

\item{node.to.extract}{\sQuote{CNVgenotype} or  \sQuote{CNVgenotypeSNPlike}. Default is  \sQuote{CNVgenotype}}
}
\value{
CNV genotypes in a specific probe
}
\description{
Extract the CNV genotypes from the GDS file
}
\author{
Vinicius Henrique da Silva <vinicius.dasilva@wur.nl>
}
