% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.comparison.R
\name{test.population}
\alias{test.population}
\title{HELPER - compare CNV segment frequencies - t.test - INCOMPLETE - Implement strategy from Maulik's paper
HELPER - Gene enrichement}
\usage{
test.population(x)
}
\description{
HELPER - compare CNV segment frequencies - t.test - INCOMPLETE - Implement strategy from Maulik's paper
HELPER - Gene enrichement
}
