% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_assoc.R
\name{testit}
\alias{testit}
\title{HELPER Wait x seconds
# from https://stat.ethz.ch/R-manual/R-devel/library/base/html/Sys.sleep.html}
\usage{
testit(x)
}
\description{
HELPER Wait x seconds
# from https://stat.ethz.ch/R-manual/R-devel/library/base/html/Sys.sleep.html
}
