% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_representations.R
\name{violinAllele}
\alias{violinAllele}
\title{Violin plot to show the quantitative phenotype distribution for each CNV genotype}
\usage{
violinAllele(all.paths, snp.name, lo.phe = 1, ylab = NULL, plot.pdf = TRUE)
}
\arguments{
\item{all.paths}{phen.info$all.paths}

\item{snp.name}{SNP name. "Name" returned by \code{\link{cnvGWAS}}}
}
\description{
Violin plot to show the quantitative phenotype distribution for each CNV genotype
}
\examples{
# Write the pdf plot in disk
 all.paths <- phen.info$all.paths
 snp.name <- "AX-76818525"
 violinAllele(all.paths, snp.name)
 
}
\author{
Vinicius Henrique da Silva <vinicius.dasilva@wur.nl>
}
