% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_representations.R
\name{plotManhattan}
\alias{plotManhattan}
\title{Manhattan Plot}
\usage{
plotManhattan(all.paths, regions, chr.size.order = NULL, plot.pdf = TRUE)
}
\arguments{
\item{all.paths}{TODO.}

\item{regions}{\code{\linkS4class{GRanges}} as returned by \code{\link{cnvGWAS}}}

\item{chr.size.order}{\code{\link{data.frame}} with two columns: (i) 'chr': chromosome names (character),
and (ii) 'size': length of the chromosomes in bp (integer)}

\item{plot.pdf}{Logical plot a to pdf file}
}
\description{
Manhattan plot for p-values of a CNV-GWAS
}
\examples{

# Load phenotype-CNV information
data.dir <- system.file("extdata", package="cnvAnalyzeR")

phen.loc <- file.path(data.dir, "Pheno.txt")
cnv.out.loc <- file.path(data.dir, "CNVOut.txt")
map.loc <- file.path(data.dir, "MapPenn.txt")

phen.info <- setupCnvGWAS('Example', phen.loc, cnv.out.loc, map.loc)
all.paths <- phen.info$all.paths
segs.pvalue.gr <- cnvGWAS(phen.info)

# Define the chromosome order in the plot
order.chrs <- c(1:24, "25LG1", "25LG2", 26:28, "LGE22", "1A", "4A")

# Chromosome sizes
chr.sizes <- c(  114059860, 150265477, 111636321, 68030631, 61875929, 34342011,
                 37653027, 31324166, 25106277, 20202851, 20315886, 20466350,
                 16480340, 16193477, 13820886, 10486032, 11572643, 9871655,
                 14661763, 7693166, 4276343, 6655392, 6808513, 1092960, 809223,
                 6596997, 4327975, 5101010, 773534, 71365269, rep(19934923,2))

chr.size.order <- data.frame(chr=order.chrs, sizes=chr.sizes, stringsAsFactors=FALSE)

# Plot Manhatthan to a pdf within the 'Results' workfolder
plotManhattan(all.paths, segs.pvalue.gr, chr.size.order)

}
\author{
Vinicius Henrique da Silva <vinicius.dasilva@wur.nl>
}
